requirements.txt
google-genai==0.3.0
aiogram==3.17.0
aiohttp==3.10.5
